pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iterm-tmuxify"
version = "0.1.0"
description = "Native iTerm panes with tmux session persistence"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
iterm-tmuxify = "iterm_tmuxify.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/iterm_tmuxify"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
